# common-platform-accounting-gl-interfaces/src/main/java/com/firefly/core/accounting/intefaces/dtos/journal/v1/JournalBatchDTO.java
/*
 * Copyright 2025 Firefly Software Solutions Inc
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */


package com.firefly.core.accounting.intefaces.dtos.journal.v1;

import com.firefly.core.accounting.intefaces.enums.journal.v1.JournalBatchStatusEnum;
import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.*;

import java.time.LocalDate;
import java.time.LocalDateTime;
import java.util.UUID;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class JournalBatchDTO {
    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
    private UUID id;

    private String reference;
    private JournalBatchStatusEnum status;
    private LocalDate batchDate;
    private String remarks;
    private LocalDateTime createdAt;
    private LocalDateTime updatedAt;
}