# common-platform-accounting-gl-interfaces/src/main/java/com/firefly/core/accounting/intefaces/dtos/journal/v1/JournalEntryDTO.java
/*
 * Copyright 2025 Firefly Software Solutions Inc
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */


package com.firefly.core.accounting.intefaces.dtos.journal.v1;

import com.firefly.core.utils.annotations.FilterableId;
import lombok.*;

import java.math.BigDecimal;
import java.time.LocalDateTime;
import java.util.UUID;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class JournalEntryDTO {
    private UUID id;

    @FilterableId
    private UUID batchId;

    @FilterableId
    private String transactionId;

    private String sourceModule;
    private BigDecimal totalAmount;
    private String description;
    private LocalDateTime entryDate;
    private boolean posted;
    private String createdBy;
    private LocalDateTime createdAt;
    private LocalDateTime updatedAt;
}